New Wall Models + Lowered Gravity (Needs Tweaks)
New Cobblestone Wall Model but requires color tweaks to match the terrain and general aesthetics.
Lowered global gravity and added a downwards force on player to act as a fake gravity only on player. Jump height needs to be decreased, move speed needs to be adjusted. Different move speed per level needs to be implemented.